refactor: add unification proxy
